problem solved! It was the number of trees that went into the calculation of lambda.... this behaves very different in the initial state compared to a simulation 50+ years down the road. Now the # of trees is calculated as the basal area  weighted number of trees.

+++ src/plugins/wind/windmodule.h
@@ -34,11 +34,11 @@ class ResourceUnit; // forward
 class WindCell {
 public:
     WindCell() { topex = 0; n_affected=0; edge_age=0; clear(); }
-    void clear() {height = edge = 0.f; n_trees=0; tree=0; n_killed = 0; basal_area_killed = 0.f; cws_uproot = 0.; cws_break= crown_windspeed= 0.; n_iteration = 0;}
+    void clear() {height = edge = 0.f; n_trees=0.f; tree=0; n_killed = 0; basal_area_killed = 0.f; cws_uproot = 0.; cws_break= crown_windspeed= 0.; n_iteration = 0;}
     bool isValid() const { return height<9999.f; } ///< returns true if the pixel is on the valid project area
     float topex; ///< topographic modifier for wind speed (-)
     float height; ///< top height (m).
-    int n_trees; ///< number of trees on pixel
+    float n_trees; ///< number of trees on pixel
     const Tree *tree; ///< pointer to the tallest tree on the pixel (if already populated)
     float edge; ///< maximum difference to neighboring cells (m)
     // statistics
@@ -125,7 +125,7 @@ private:
     /// perform the wind effect calculations for a given grid cell
     bool windImpactOnPixel(const QPoint position, WindCell *cell);
     ///
-    double calculateCrownWindSpeed(const Tree *tree, const WindSpeciesParameters &params, const int n_trees, const double wind_speed_10);
+    double calculateCrownWindSpeed(const Tree *tree, const WindSpeciesParameters &params, const double n_trees, const double wind_speed_10);
     double calculateCrititalWindSpeed(const Tree *tree, const WindSpeciesParameters &params, const double gap_length, double &rCWS_uproot, double &rCWS_break);
     /// check if the soil on current resource unit is frozen or not
     bool isSoilFrozen(const ResourceUnit *ru, const int day_of_year) const;
